Refactor with MetricMixin
