+++ api/providers/views.py
@@ -6,9 +6,10 @@ from rest_framework import permissions as drf_permissions
 from rest_framework.exceptions import NotAuthenticated, NotFound
 
 from api.base import permissions as base_permissions
-from api.base.exceptions import InvalidFilterValue, InvalidFilterOperator, Conflict, InvalidQueryStringError
+from api.base.exceptions import InvalidFilterValue, InvalidFilterOperator, Conflict
 from api.base.filters import PreprintFilterMixin, ListFilterMixin
 from api.base.views import JSONAPIBaseView
+from api.base.metrics import MetricMixin
 from api.base.pagination import MaxSizePagination, IncreasedPageSizePagination
 from api.base.utils import get_object_or_error, get_user_auth, is_truthy
 from api.licenses.views import LicenseList
@@ -62,7 +63,7 @@ class RegistrationProviderList(GenericProviderList):
     view_name = 'registration-providers-list'
 
 
-class PreprintProviderList(GenericProviderList):
+class PreprintProviderList(MetricMixin, GenericProviderList):
     """The documentation for this endpoint can be found [here](https://developer.osf.io/#operation/preprint_provider_list).
     """
 
@@ -70,26 +71,17 @@ class PreprintProviderList(GenericProviderList):
     serializer_class = PreprintProviderSerializer
     view_category = 'preprint-providers'
     view_name = 'preprint-providers-list'
+    metric_map = {
+        'downloads': PreprintDownload
+    }
 
-    def get_default_queryset(self):
-        METRIC_MAP = {
-            'downloads': PreprintDownload
-        }
-        queryset = super(PreprintProviderList, self).get_default_queryset()
-        metric_param = self.request.query_params.get('metrics', None)
-        if metric_param:
-            metrics = [each.lower().strip() for each in metric_param.split(',')]
-            for metric in metrics:
-                if metric not in METRIC_MAP:
-                    raise InvalidQueryStringError('Invalid metric in query string: {}'.format(metric))
-                metric_class = METRIC_MAP[metric]
-                queryset = metric_class.get_top_by_count(
-                    qs=queryset,
-                    model_field='_id',
-                    metric_field='provider_id',
-                    annotation='downloads',
-                )
-        return queryset
+    def get_annotated_queryset_with_metrics(self, queryset, metric_class, metric_name):
+        return metric_class.get_top_by_count(
+            qs=queryset,
+            model_field='_id',
+            metric_field='provider_id',
+            annotation=metric_name,
+        )
 
     def get_renderer_context(self):
         context = super(PreprintProviderList, self).get_renderer_context()
